backend/requirements.txt
# Backend requirements for Style Vibe Classifier
fastapi>=0.136.0
uvicorn[standard]>=0.46.0
python-multipart>=0.0.20
torch>=2.11.0
torchvision>=0.26.0
transformers>=5.6.0
Pillow>=12.2.0
scikit-learn
tqdm

# Model training
google-generativeai
pandas
datasets
pillow
python-dotenv

# Model visualization w/ Jupyter notebook
jupyter
matplotlib
seaborn